+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 #### :rocket: New Feature
 
 - Allow `@directive` on functions for emitting function level directive code (`let serverAction = @directive("'use server'") (~name) => {...}`). https://github.com/rescript-lang/rescript-compiler/pull/6756
+- Add new dict literal syntax (`dict{"foo": "bar"}`). https://github.com/rescript-lang/rescript-compiler/pull/6774
 
 #### :boom: Breaking Change